+++ Controller.php
@@ -7,7 +7,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Plugins\LoginOIDC;
+namespace Piwik\Plugins\PkceOIDC;
 
 use Exception;
 use Piwik\Access;
@@ -33,7 +33,7 @@ class Controller extends \Piwik\Plugin\Controller
      *
      * @var string
      */
-    const OIDC_NONCE = "LoginOIDC.nonce";
+    const OIDC_NONCE = "PkceOIDC.nonce";
 
     /**
      * Auth implementation to login users.
@@ -68,7 +68,7 @@ class Controller extends \Piwik\Plugin\Controller
         parent::__construct();
 
         if (empty($auth)) {
-            $auth = StaticContainer::get("Piwik\Plugins\LoginOIDC\Auth");
+            $auth = StaticContainer::get("Piwik\Plugins\PkceOIDC\Auth");
         }
         $this->auth = $auth;
 
@@ -105,7 +105,7 @@ class Controller extends \Piwik\Plugin\Controller
      */
     public function loginMod() : string
     {
-        $settings = new \Piwik\Plugins\LoginOIDC\SystemSettings();
+        $settings = new \Piwik\Plugins\PkceOIDC\SystemSettings();
         return $this->renderTemplate("loginMod", array(
             "caption" => $settings->authenticationName->getValue(),
             "nonce" => Nonce::getNonce(self::OIDC_NONCE)
@@ -131,7 +131,7 @@ class Controller extends \Piwik\Plugin\Controller
     public function unlink()
     {
         if ($_SERVER["REQUEST_METHOD"] !== "POST") {
-            throw new Exception(Piwik::translate("LoginOIDC_MethodNotAllowed"));
+            throw new Exception(Piwik::translate("PkceOIDC_MethodNotAllowed"));
         }
         // csrf protection
         Nonce::checkNonce(self::OIDC_NONCE, $_POST["form_nonce"]);
@@ -149,14 +149,14 @@ class Controller extends \Piwik\Plugin\Controller
      */
     public function signin()
     {
-        $settings = new \Piwik\Plugins\LoginOIDC\SystemSettings();
+        $settings = new \Piwik\Plugins\PkceOIDC\SystemSettings();
 
         $allowedMethods = array("POST");
         if (!$settings->disableDirectLoginUrl->getValue()) {
             array_push($allowedMethods, "GET");
         }
         if (!in_array($_SERVER["REQUEST_METHOD"], $allowedMethods)) {
-            throw new Exception(Piwik::translate("LoginOIDC_MethodNotAllowed"));
+            throw new Exception(Piwik::translate("PkceOIDC_MethodNotAllowed"));
         }
 
         if ($_SERVER["REQUEST_METHOD"] === "POST") {
@@ -165,19 +165,29 @@ class Controller extends \Piwik\Plugin\Controller
         }
 
         if (!$this->isPluginSetup($settings)) {
-            throw new Exception(Piwik::translate("LoginOIDC_ExceptionNotConfigured"));
+            throw new Exception(Piwik::translate("PkceOIDC_ExceptionNotConfigured"));
         }
 
         $_SESSION["loginoidc_state"] = $this->generateKey(32);
+        $codeVerifier = $this->generateKey(64);
+        $_SESSION['loginoidc_pkce_verifier'] = $codeVerifier;
+        $codeChallenge = $this->base64UrlEncode(hash('sha256', $codeVerifier, true));
         $params = array(
             "client_id" => $settings->clientId->getValue(),
             "scope" => $settings->scope->getValue(),
-            "redirect_uri"=> $this->getRedirectUri(),
+            "redirect_uri" => $this->getRedirectUri(),
             "state" => $_SESSION["loginoidc_state"],
-            "response_type" => "code"
+            "response_type" => "code",
+            "code_challenge" => $codeChallenge,
+            "code_challenge_method" => "S256"
         );
+
+        // Build query string without redirect_uri
+        $queryString = http_build_query($params);
+
         $url = $settings->authorizeUrl->getValue();
-        $url .= (parse_url($url, PHP_URL_QUERY) ? "&" : "?") . http_build_query($params);
+        $url .= (parse_url($url, PHP_URL_QUERY) ? "&" : "?") . $queryString;
+
         Url::redirectToUrl($url);
     }
 
@@ -189,19 +199,22 @@ class Controller extends \Piwik\Plugin\Controller
      */
     public function callback()
     {
-        $settings = new \Piwik\Plugins\LoginOIDC\SystemSettings();
+        if (!ob_get_level()) {
+            ob_start();
+        }
+        $settings = new \Piwik\Plugins\PkceOIDC\SystemSettings();
         if (!$this->isPluginSetup($settings)) {
-            throw new Exception(Piwik::translate("LoginOIDC_ExceptionNotConfigured"));
+            throw new Exception(Piwik::translate("PkceOIDC_ExceptionNotConfigured"));
         }
 
         if ($_SESSION["loginoidc_state"] !== Request::fromGet()->getStringParameter("state")) {
-            throw new Exception(Piwik::translate("LoginOIDC_ExceptionStateMismatch"));
+            throw new Exception(Piwik::translate("PkceOIDC_ExceptionStateMismatch"));
         } else {
             unset($_SESSION["loginoidc_state"]);
         }
 
         if (Request::fromGet()->getStringParameter("provider") !== "oidc") {
-            throw new Exception(Piwik::translate("LoginOIDC_ExceptionUnknownProvider"));
+            throw new Exception(Piwik::translate("PkceOIDC_ExceptionUnknownProvider"));
         }
 
         // payload for token request
@@ -213,6 +226,10 @@ class Controller extends \Piwik\Plugin\Controller
             "grant_type" => "authorization_code",
             "state" => Request::fromGet()->getStringParameter("state")
         );
+        if (isset($_SESSION['loginoidc_pkce_verifier'])) {
+            $data['code_verifier'] = $_SESSION['loginoidc_pkce_verifier'];
+            unset($_SESSION['loginoidc_pkce_verifier']);
+        }
         $dataString = http_build_query($data);
 
         $curl = curl_init();
@@ -222,7 +239,7 @@ class Controller extends \Piwik\Plugin\Controller
             "Content-Type: application/x-www-form-urlencoded",
             "Content-Length: " . strlen($dataString),
             "Accept: application/json",
-            "User-Agent: LoginOIDC-Matomo-Plugin"
+            "User-Agent: PkceOIDC-Matomo-Plugin"
         ));
         curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
         curl_setopt($curl, CURLOPT_URL, $settings->tokenUrl->getValue());
@@ -232,7 +249,7 @@ class Controller extends \Piwik\Plugin\Controller
         $result = json_decode($response);
 
         if (empty($result) || empty($result->access_token)) {
-            throw new Exception(Piwik::translate("LoginOIDC_ExceptionInvalidResponse"));
+            throw new Exception(Piwik::translate("PkceOIDC_ExceptionInvalidResponse"));
         }
 
         $_SESSION['loginoidc_idtoken'] = empty($result->id_token) ? null : $result->id_token;
@@ -242,7 +259,7 @@ class Controller extends \Piwik\Plugin\Controller
         curl_setopt($curl, CURLOPT_HTTPHEADER, array(
             "Authorization: Bearer " . $result->access_token,
             "Accept: application/json",
-            "User-Agent: LoginOIDC-Matomo-Plugin"
+            "User-Agent: PkceOIDC-Matomo-Plugin"
         ));
         curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
         curl_setopt($curl, CURLOPT_URL, $settings->userinfoUrl->getValue());
@@ -255,7 +272,7 @@ class Controller extends \Piwik\Plugin\Controller
         $providerUserId = $result->$userinfoId;
 
         if (empty($providerUserId)) {
-            throw new Exception(Piwik::translate("LoginOIDC_ExceptionInvalidResponse"));
+            throw new Exception(Piwik::translate("PkceOIDC_ExceptionInvalidResponse"));
         }
 
         $user = $this->getUserByRemoteId("oidc", $providerUserId);
@@ -286,7 +303,7 @@ class Controller extends \Piwik\Plugin\Controller
             // users identity has been successfully confirmed by the remote oidc server
             if (Piwik::isUserIsAnonymous()) {
                 if ($settings->disableSuperuser->getValue() && $this->hasTheUserSuperUserAccess($user["login"])) {
-                    throw new Exception(Piwik::translate("LoginOIDC_ExceptionSuperUserOauthDisabled"));
+                    throw new Exception(Piwik::translate("PkceOIDC_ExceptionSuperUserOauthDisabled"));
                 } else {
                     $this->signinAndRedirect($user, $settings);
                 }
@@ -295,10 +312,15 @@ class Controller extends \Piwik\Plugin\Controller
                     $this->passwordVerify->setPasswordVerifiedCorrectly();
                     return;
                 } else {
-                    throw new Exception(Piwik::translate("LoginOIDC_ExceptionAlreadyLinkedToDifferentAccount"));
+                    throw new Exception(Piwik::translate("PkceOIDC_ExceptionAlreadyLinkedToDifferentAccount"));
                 }
             }
         }
+    }
+
+    private function base64UrlEncode($data)
+    {
+        return rtrim(strtr(base64_encode($data), '+/', '-_'), '=');
     }
 
     /**
@@ -307,7 +329,7 @@ class Controller extends \Piwik\Plugin\Controller
      * It was used as a template for this function.
      * See: {@link \Piwik\Core::hasTheUserSuperUserAccess($theUser)} method.
      * See: {@link \Piwik\Plugins\UsersManager\Model::getUsersHavingSuperUserAccess()} method.
-     * 
+     *
      * @param  string  $theUser A username to be checked for superuser access
      * @return bool
      */
@@ -371,7 +393,7 @@ class Controller extends \Piwik\Plugin\Controller
         if ($settings->allowSignup->getValue()) {
             // verify response contains email address
             if (empty($matomoUserLogin)) {
-                throw new Exception(Piwik::translate("LoginOIDC_ExceptionUserNotFoundAndNoEmail"));
+                throw new Exception(Piwik::translate("PkceOIDC_ExceptionUserNotFoundAndNoEmail"));
             }
 
             // verify email address domain is allowed to sign up
@@ -379,7 +401,7 @@ class Controller extends \Piwik\Plugin\Controller
                 $signupDomain = substr($matomoUserLogin, strpos($matomoUserLogin, "@") + 1);
                 $allowedDomains = explode("\n", $settings->allowedSignupDomains->getValue());
                 if (!in_array($signupDomain, $allowedDomains)) {
-                    throw new Exception(Piwik::translate("LoginOIDC_ExceptionAllowedSignupDomainsDenied"));
+                    throw new Exception(Piwik::translate("PkceOIDC_ExceptionAllowedSignupDomainsDenied"));
                 }
             }
 
@@ -396,7 +418,7 @@ class Controller extends \Piwik\Plugin\Controller
             $this->linkAccount($providerUserId, $matomoUserLogin);
             $this->signinAndRedirect($user, $settings);
         } else {
-            throw new Exception(Piwik::translate("LoginOIDC_ExceptionUserNotFoundAndSignupDisabled"));
+            throw new Exception(Piwik::translate("PkceOIDC_ExceptionUserNotFoundAndSignupDisabled"));
         }
     }
 
@@ -411,6 +433,9 @@ class Controller extends \Piwik\Plugin\Controller
         $this->auth->setLogin($user["login"]);
         $this->auth->setForceLogin(true);
         $this->sessionInitializer->initSession($this->auth);
+        if (ob_get_level()) {
+            ob_end_flush();
+        }
         if ($settings->bypassTwoFa->getValue()) {
             $sessionFingerprint = new SessionFingerprint();
             $sessionFingerprint->setTwoFactorAuthenticationVerified();
@@ -439,13 +464,13 @@ class Controller extends \Piwik\Plugin\Controller
      */
     private function getRedirectUri() : string
     {
-        $settings = new \Piwik\Plugins\LoginOIDC\SystemSettings();
+        $settings = new \Piwik\Plugins\PkceOIDC\SystemSettings();
 
         if (!empty($settings->redirectUriOverride->getValue())) {
             return $settings->redirectUriOverride->getValue();
         } else {
             $params = array(
-                "module" => "LoginOIDC",
+                "module" => "PkceOIDC",
                 "action" => "callback",
                 "provider" => "oidc"
             );